add shell command support

+++ jasmine/src/main/java/org/yidan/idea/plugin/jasmine/Logger.java
@@ -11,6 +11,6 @@ public interface Logger {
 
     void info(String message);
 
-    void setProgress(double percent);
+    void setProgress(int percent);
 
 }
